requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
google-auth-oauthlib==0.4.6
google-auth==2.28.1
pydantic==2.6.3
